Fix ExtractKeyIds tuple ordering and clean up
Reorder ExtractKeyIds return tuple to (requestId, operationId) with
explicit named elements and W3C format comments. Update all call sites
for consistency.
