+++ src/ServiceProfiler.EventPipe.Otel/Azure.Monitor.OpenTelemetry.Profiler.Core/EventListeners/TraceSessionListenerFactory.cs
@@ -5,13 +5,6 @@ namespace Azure.Monitor.OpenTelemetry.Profiler.Core.EventListeners;
 
 internal class TraceSessionListenerFactory
 {
-    // Internal, undocumented kill switch. Lets us A/B the two request-event sources during the
-    // migration away from OpenTelemetry-Sdk's RequestStart/Stop toward DiagnosticSource.
-    // Accepted values (case-insensitive): "ds" | "otel" | "both". Anything else falls back to default.
-    internal const string RequestSourceEnvVarName = "MICROSOFT_PROFILER_INTERNAL_REQUEST_SOURCE";
-
-    private const RequestSourceMode DefaultRequestSourceMode = RequestSourceMode.DiagnosticSource;
-
     private readonly IServiceProvider _serviceProvider;
 
     public TraceSessionListenerFactory(IServiceProvider serviceProvider)
@@ -24,7 +17,7 @@ internal class TraceSessionListenerFactory
         ILogger<TraceSessionListenerFactory> logger = _serviceProvider
             .GetRequiredService<ILogger<TraceSessionListenerFactory>>();
 
-        RequestSourceMode mode = ReadRequestSourceMode(logger);
+        RequestSourceMode mode = RequestSourceModeResolver.Resolve(logger);
         logger.LogInformation("Request event source mode: {mode}", mode);
 
         // One RequestActivityRelay per listener lifetime, shared across this listener's handlers so that
@@ -47,30 +40,5 @@ internal class TraceSessionListenerFactory
         handlers.Add(ActivatorUtilities.CreateInstance<TplEventSourceHandler>(_serviceProvider));
 
         return ActivatorUtilities.CreateInstance<TraceSessionListener>(_serviceProvider, (IEnumerable<IEventSourceHandler>)handlers);
-    }
-
-    private static RequestSourceMode ReadRequestSourceMode(ILogger logger)
-    {
-        string? raw = Environment.GetEnvironmentVariable(RequestSourceEnvVarName);
-        if (string.IsNullOrWhiteSpace(raw))
-        {
-            return DefaultRequestSourceMode;
-        }
-
-        return raw.Trim().ToLowerInvariant() switch
-        {
-            "ds" => RequestSourceMode.DiagnosticSource,
-            "otel" => RequestSourceMode.OpenTelemetrySdk,
-            "both" => RequestSourceMode.Both,
-            _ => LogUnrecognizedAndUseDefault(logger, raw),
-        };
-    }
-
-    private static RequestSourceMode LogUnrecognizedAndUseDefault(ILogger logger, string raw)
-    {
-        logger.LogWarning(
-            "Unrecognized value '{value}' for {envVar}; falling back to {default}. Expected: ds | otel | both.",
-            raw, RequestSourceEnvVarName, DefaultRequestSourceMode);
-        return DefaultRequestSourceMode;
     }
 }
